refactor(api): rename files with old mention to schooling registration
